+++ src/IEventSubscriber.php
@@ -3,10 +3,9 @@ declare(strict_types=1);
 
 namespace Konecnyjakub\EventDispatcher;
 
-interface IEventSubscriber
+/**
+ * @deprecated Use {@see EventSubscriber} instead
+ */
+interface IEventSubscriber extends EventSubscriber
 {
-    /**
-     * @return iterable<class-string, array{0: string, 1?: int}[]>
-     */
-    public static function getSubscribedEvents(): iterable;
 }